Improvements in build script. Corrections test classes names
