+++ build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id "org.jetbrains.kotlin.jvm" version "1.3.30"
+    id "org.jetbrains.kotlin.jvm" version "1.3.31"
     id "maven-publish"
 }
 
@@ -7,25 +7,24 @@ group = 'org.openbroker'
 version = '1.0-SNAPSHOT'
 description = 'Library for Open Broker'
 
-defaultTasks 'test'
-
 ext {
     kotlinLanguageVersion = '1.3'
+    jdkVersion = JavaVersion.VERSION_1_8
 
     defaultEncoding = 'UTF-8'
 
     ver = [
-        'kotlin'     : '1.3.30',
-        'springBoot' : '2.1.4.RELEASE',
-        'cloudEvents': '76dfa0814f',
+        'kotlin'     : '1.3.31',
+        'jackson'    : '2.9.9',
+        'junit'      : '5.4.2',
+        'cloudEvents': 'b8b9c6fef3',
     ]
 
     servicePackage = 'org.openbroker'
-    servicePackageSlashes = servicePackage.replace('.', '/')
 }
 
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
+sourceCompatibility = jdkVersion
+targetCompatibility = jdkVersion
 
 repositories {
     jcenter()
@@ -36,16 +35,17 @@ repositories {
 dependencies {
 
     implementation(
-        platform("org.springframework.boot:spring-boot-dependencies:$ver.springBoot"),
         platform("org.jetbrains.kotlin:kotlin-bom:$ver.kotlin"),
+        platform("com.fasterxml.jackson:jackson-bom:$ver.jackson"),
         "org.jetbrains.kotlin:kotlin-stdlib-jdk8",
-        "com.github.open-broker:cloud-events-jvm:$ver.cloudEvents",
         "com.fasterxml.jackson.core:jackson-core",
         "com.fasterxml.jackson.module:jackson-module-kotlin",
         "com.fasterxml.jackson.datatype:jackson-datatype-jsr310",
+        "com.github.open-broker:cloud-events-jvm:$ver.cloudEvents",
     )
 
     testImplementation(
+        platform("org.junit:junit-bom:$ver.junit"),
         "org.junit.jupiter:junit-jupiter-api",
     )
 
@@ -57,7 +57,7 @@ dependencies {
 tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
     kotlinOptions {
         javaParameters = true
-        jvmTarget = JavaVersion.VERSION_1_8
+        jvmTarget = jdkVersion
         freeCompilerArgs += ['-Xjsr305=strict', '-progressive']
         languageVersion = kotlinLanguageVersion
         apiVersion = kotlinLanguageVersion
@@ -91,7 +91,7 @@ test {
     reports.html.enabled false
 
     failFast = System.getProperty('failFast') != null
-    include servicePackageSlashes + '/**/*Test.*'
+    include servicePackage.replace('.', '/') + '/**/*Test.*'
 
     testLogging {
         events 'skipped', 'failed'
