+++ gradle.properties
@@ -1,8 +1,4 @@
-# The Gradle daemon aims to improve the startup and execution time of Gradle.
-# When set to true the Gradle daemon is to run the build.
 org.gradle.daemon=false
-# Specifies the JVM arguments used for the daemon process.
-# The setting is particularly useful for tweaking memory settings.
-# Default value: -Xmx1024m -XX:MaxPermSize=256m
-org.gradle.jvmargs=-Xmx1024m -XX:MaxPermSize=256m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
 org.gradle.caching=true
+org.gradle.console=plain
+org.gradle.jvmargs=-XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
