+++ src/test/java/org/openbroker/no/serialize/SerializationJavaTest.java
@@ -1,20 +1,21 @@
-package org.openbroker.se.serialize;
+package org.openbroker.no.serialize;
 
 import org.junit.jupiter.api.Test;
 import org.openbroker.cloudevents.CloudEvent;
-import org.openbroker.se.privateunsecuredloan.TestObjectsJson;
 import org.openbroker.cloudevents.SerializationKt;
-import org.openbroker.se.privateunsecuredloan.events.ApplicationCreated;
+import org.openbroker.no.privateunsecuredloan.TestObjectsJson;
+import org.openbroker.no.privateunsecuredloan.events.ApplicationCreated;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class SerializationTestJava {
+class SerializationJavaTest {
+
     @Test
-    public void testDeserializeOpenBrokerApplicationCreatedToType() {
+    void testDeserializeOpenBrokerApplicationCreatedToType() {
         CloudEvent<ApplicationCreated> event = SerializationKt.cloudEvent(TestObjectsJson.fullApplicationCreatedJson, ApplicationCreated.class);
         assertNotNull(event.getData());
         ApplicationCreated app = event.getData();
-        assertEquals(app.getApplication().getApplicant().getTentativeAddress().getFirstName(), "Christin");
+        assertEquals(app.getApplication().getApplicant().getTentativeAddress().getFirstName(), "Jane");
     }
 }
