+++ src/test/java/org/openbroker/se/OpenBrokerCreateTest.java
@@ -9,10 +9,10 @@ import org.openbroker.se.privateunsecuredloan.events.OfferAccepted;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class OpenBrokerCreateTest {
+class OpenBrokerCreateTest {
 
     @Test
-    public void testCreateOpenBrokerEvent() {
+    void testCreateOpenBrokerEvent() {
         CloudEvent<OfferAccepted> accept = OpenBrokerEventKt.create(
             new OfferAccepted(new Reference("1", "io.klira"), new BankAccount("3300", "1234567890")),
             OfferAccepted.class,
