+++ src/test/java/org/openbroker/se/privateunsecuredloan/EventTypeFactorTest.java
@@ -10,7 +10,7 @@ import java.util.stream.Collectors;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class EventTypeFactor {
+class EventTypeFactorTest {
 
     @Test
     void testCorrectSortingByComparatorWithStatusUpdatedOutOfOrder() {
