+++ src/test/kotlin/org/openbroker/no/privateunsecuredloan/EventTypeFactoryTest.kt
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.openbroker.no.privateunsecuredloan.events.PrivateUnsecuredLoanEvent
 
-class EventTypeFactory {
+class EventTypeFactoryTest {
 
     @Test
     fun testCorrectSortingByComparator() {
